use metricsender in external file
