+++ oauthproxy.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/18F/hmacauth"
 	"github.com/alexcesaro/statsd"
-	"github.com/rainycape/unidecode"
 	"github.com/skbkontur/oauth2_proxy/cookie"
 	"github.com/skbkontur/oauth2_proxy/providers"
 )
@@ -729,28 +728,7 @@ func (p *OAuthProxy) CheckBasicAuth(req *http.Request) (*providers.SessionState,
 	}
 	if p.HtpasswdFile.Validate(pair[0], pair[1]) {
 		log.Printf("authenticated %q via basic auth", pair[0])
-		return &providers.SessionState{User: pair[0], BasicAuth: true}, nil
+		return &providers.SessionState{User: pair[0]}, nil
 	}
 	return nil, fmt.Errorf("%s not in HtpasswdFile", pair[0])
-}
-
-func (p *OAuthProxy) incrementAuthenticated(session *providers.SessionState, method string) {
-	if session.BasicAuth == true {
-		userAlias := unidecode.Unidecode(session.User)
-		userAlias = strings.Replace(userAlias, ".", "-", -1)
-		metricName := fmt.Sprintf("authenticated.BasicAuth.%s.%s", userAlias, method)
-		p.StatsD.Increment(metricName)
-	} else {
-		for _, group := range session.Groups {
-			groupAlias := unidecode.Unidecode(group)
-			groupAlias = strings.Replace(groupAlias, ".", "-", -1)
-			metricName := fmt.Sprintf("authenticated.Oauth2.%s.%s", groupAlias, method)
-			p.StatsD.Increment(metricName)
-		}
-	}
-}
-
-func (p *OAuthProxy) incrementUnauthenticated(method string, returnedStatus string) {
-	metricName := fmt.Sprintf("unauthenticated.%s.resultedWith.%s", method, returnedStatus)
-	p.StatsD.Increment(metricName)
 }
