+++ providers/session_state.go
@@ -16,7 +16,6 @@ type SessionState struct {
 	Email        string
 	User         string
 	Groups       []string
-	BasicAuth    bool
 }
 
 func (s *SessionState) IsExpired() bool {
